Got the retron 5 cart erasing working
