+++ frontend/src/components/Retron5EraseCartridge.vue
@@ -10,52 +10,19 @@
               </b-jumbotron>
             </b-col>
           </b-row>
-          <div v-if="this.conversionDirection === 'convertToEmulator'">
-            <input-file
-              @load="readRetron5SaveData($event)"
-              :errorMessage="this.errorMessage"
-              placeholderText="Choose a file to convert (*.sav)"
-              acceptExtension=".sav"
-              :leaveRoomForHelpIcon="false"
-            />
-          </div>
-          <div v-else>
-            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="false"/>
-          </div>
-        </b-col>
-        <b-col sm=12 md=2 lg=2 xl=2 align-self="start">
-          <conversion-direction
-            :horizontalLayout="['md', 'lg', 'xl']"
-            :verticalLayout="['xs', 'sm']"
-            :conversionDirection="this.conversionDirection"
-            @change="changeConversionDirection($event)"
+          <input-file
+            @load="readRetron5SaveData($event)"
+            :errorMessage="this.errorMessage"
+            placeholderText="Choose a file to convert (*.sav)"
+            acceptExtension=".sav"
+            :leaveRoomForHelpIcon="false"
           />
-        </b-col>
-        <b-col sm=12 md=5 align-self="start">
-          <b-row no-gutters align-h="center" align-v="start">
-            <b-col cols=12>
-              <b-jumbotron fluid :header-level="$mq | mq({ xs: 5, sm: 5, md: 5, lg: 5, xl: 4 })">
-                <template v-slot:header>Emulator/Raw</template>
-              </b-jumbotron>
-            </b-col>
-          </b-row>
-          <div v-if="this.conversionDirection === 'convertToEmulator'">
-            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="false"/>
-          </div>
-          <div v-else>
-            <input-file
-              @load="readEmulatorSaveData($event)"
-              :errorMessage="this.errorMessage"
-              placeholderText="Choose a file to convert"
-              :leaveRoomForHelpIcon="false"
-            />
-          </div>
         </b-col>
       </b-row>
       <b-row class="justify-content-md-center" align-h="center">
         <b-col cols="auto" sm=4 md=3 lg=2 align-self="center">
           <b-button
-            class="retron5-convert-button"
+            class="retron5-erase-cartridge-convert-button"
             variant="success"
             block
             :disabled="!this.retron5SaveData || !outputFilename"
@@ -79,7 +46,7 @@
 <style scoped>
 
 /* Separate class for each different button to enable tracking in google tag manager */
-.retron5-convert-button {
+.retron5-erase-cartridge-convert-button {
   margin-top: 1em;
 }
 
@@ -90,10 +57,9 @@
 
 <script>
 import { saveAs } from 'file-saver';
+import SaveFilesUtil from '../util/SaveFiles';
 import Util from '../util/util';
 import InputFile from './InputFile.vue';
-import OutputFilename from './OutputFilename.vue';
-import ConversionDirection from './ConversionDirection.vue';
 import Retron5SaveData from '../save-formats/Retron5';
 
 export default {
@@ -103,43 +69,27 @@ export default {
       retron5SaveData: null,
       errorMessage: null,
       outputFilename: null,
-      conversionDirection: 'convertToEmulator',
     };
   },
   components: {
-    ConversionDirection,
     InputFile,
-    OutputFilename,
   },
   methods: {
-    changeConversionDirection(newDirection) {
-      this.conversionDirection = newDirection;
-      this.retron5SaveData = null;
-      this.errorMessage = null;
-      this.outputFilename = null;
-    },
     readRetron5SaveData(event) {
       this.errorMessage = null;
+      this.outputFilename = null;
       try {
         this.retron5SaveData = Retron5SaveData.createFromRetron5Data(event.arrayBuffer);
-        this.outputFilename = Util.changeFilenameExtension(event.filename, 'srm');
-      } catch (e) {
-        this.errorMessage = e.message;
-        this.retron5SaveData = null;
-      }
-    },
-    readEmulatorSaveData(event) {
-      this.errorMessage = null;
-      try {
-        this.retron5SaveData = Retron5SaveData.createFromEmulatorData(event.arrayBuffer);
-        this.outputFilename = Util.changeFilenameExtension(event.filename, 'sav');
+        this.outputFilename = Util.getFilename(event.filename);
       } catch (e) {
         this.errorMessage = e.message;
         this.retron5SaveData = null;
       }
     },
     convertFile() {
-      const outputArrayBuffer = (this.conversionDirection === 'convertToEmulator') ? this.retron5SaveData.getRawSaveData() : this.retron5SaveData.getArrayBuffer();
+      const outputRawSave = SaveFilesUtil.getEraseCartridgeSave(this.retron5SaveData.getRawSaveData());
+
+      const outputArrayBuffer = Retron5SaveData.createFromEmulatorData(outputRawSave).getArrayBuffer();
 
       const outputBlob = new Blob([outputArrayBuffer], { type: 'application/octet-stream' });
